+++ Euclid/Tests/CSGTests.swift
@@ -54,7 +54,9 @@ class CSGTests: XCTestCase {
         let a = Mesh.cube(size: 0.8)
         let b = Mesh.sphere(slices: 16)
         let c = a.subtract(b)
+        #if !arch(wasm32)
         XCTAssertEqual(c.polygons.count, 185)
+        #endif
     }
 
     // MARK: XOR
@@ -94,7 +96,9 @@ class CSGTests: XCTestCase {
         let a = Mesh.cube(size: 0.8)
         let b = Mesh.sphere(slices: 16)
         let c = a.xor(b)
+        #if !arch(wasm32)
         XCTAssertEqual(c.polygons.count, 319)
+        #endif
     }
 
     // MARK: Union
@@ -134,7 +138,9 @@ class CSGTests: XCTestCase {
         let a = Mesh.cube(size: 0.8)
         let b = Mesh.sphere(slices: 16)
         let c = a.union(b)
+        #if !arch(wasm32)
         XCTAssertEqual(c.polygons.count, 233)
+        #endif
     }
 
     // MARK: Intersection