+++ Euclid/Tests/MeshShapeTests.swift
@@ -87,7 +87,9 @@ class MeshShapeTests: XCTestCase {
         // Every vertex in the loft should be contained by one of our shapes
         let vertices = loft.polygons.flatMap { $0.vertices }
         XCTAssert(vertices.allSatisfy { vertex in
-            shapes.contains(where: { $0.points.contains(where: { $0.position == vertex.position }) })
+            shapes.contains(where: {
+                $0.points.contains(where: { $0.position == vertex.position })
+            })
         })
     }
 
@@ -105,13 +107,81 @@ class MeshShapeTests: XCTestCase {
 
         let loft = Mesh.loft(shapes)
 
-        XCTAssert(loft.polygons.allSatisfy { pointsAreCoplanar($0.vertices.map { $0.position }) })
+        XCTAssert(loft.polygons.allSatisfy {
+            pointsAreCoplanar($0.vertices.map { $0.position })
+        })
 
         // Every vertex in the loft should be contained by one of our shapes
         let vertices = loft.polygons.flatMap { $0.vertices }
         XCTAssert(vertices.allSatisfy { vertex in
-            shapes.contains(where: { $0.points.contains(where: { $0.position == vertex.position }) })
+            shapes.contains(where: {
+                $0.points.contains(where: { $0.position == vertex.position })
+            })
         })
+    }
+
+    func testLoftNonParallelEdges2() {
+        let shapes = [
+            Path.circle().rotated(by: Rotation(yaw: .pi / 8)),
+            Path.circle().rotated(by: Rotation(yaw: -.pi / 8))
+                .translated(by: Vector(0, 0, 1)),
+        ]
+
+        let loft = Mesh.loft(shapes)
+        XCTAssert(loft.isWatertight)
+        XCTAssert(loft.polygons.areWatertight)
+        XCTAssertEqual(loft.polygons.count, 18)
+    }
+
+    func testLoftCoincidentClosedPaths() {
+        let shapes = [
+            Path.square(),
+            Path.square(),
+        ]
+
+        let loft = Mesh.loft(shapes)
+        XCTAssert(loft.isWatertight)
+        XCTAssert(loft.polygons.areWatertight)
+        XCTAssertEqual(loft.polygons.count, 2)
+    }
+
+    func testLoftOffsetOpenPaths() {
+        let shapes = [
+            Path.line(Vector(0, 0), Vector(0, 1)),
+            Path.line(Vector(0, 0, 1), Vector(0, 1, 1)),
+        ]
+
+        let loft = Mesh.loft(shapes)
+        XCTAssertEqual(loft.polygons.count, 2)
+        XCTAssert(loft.polygons.areWatertight)
+    }
+
+    func testLoftCoincidentOpenPaths() {
+        let shapes = [
+            Path.line(Vector(0, 0), Vector(0, 1)),
+            Path.line(Vector(0, 0), Vector(0, 1)),
+        ]
+
+        let loft = Mesh.loft(shapes)
+        XCTAssertEqual(loft, .empty)
+    }
+
+    func testLoftClosedToOpenToClosedPath() {
+        let shapes = [
+            Path.square(),
+            Path([
+                .point(-1, -1, 1),
+                .point(1, -1, 1),
+                .point(1, 1, 1),
+                .point(-1, 1, 1),
+            ]),
+            Path.square().translated(by: Vector(0, 0, 2)),
+        ]
+
+        let loft = Mesh.loft(shapes)
+        XCTAssert(loft.isWatertight)
+        XCTAssert(loft.polygons.areWatertight)
+        XCTAssertEqual(loft.polygons.count, 20)
     }
 
     func testExtrudeSelfIntersectingPath() {
@@ -133,6 +203,8 @@ class MeshShapeTests: XCTestCase {
             .point(0, 0),
         ])
         let mesh = Mesh.extrude(path)
+        XCTAssert(mesh.isWatertight)
+        XCTAssert(mesh.polygons.areWatertight)
         XCTAssertEqual(mesh.polygons.count, 2)
         XCTAssertEqual(mesh, .extrude(path, faces: .front))
     }
@@ -143,8 +215,116 @@ class MeshShapeTests: XCTestCase {
             .point(0, 1),
         ])
         let mesh = Mesh.extrude(path)
+        XCTAssert(mesh.isWatertight)
+        XCTAssert(mesh.polygons.areWatertight)
         XCTAssertEqual(mesh.polygons.count, 2)
         XCTAssertEqual(mesh, .extrude(path, faces: .frontAndBack))
+    }
+
+    func testExtrudeOpenLineAlongClosedPath() {
+        let path = Path([
+            .point(0, 0),
+            .point(0, 1),
+        ])
+        let mesh = Mesh.extrude(path, along: .square())
+        XCTAssert(mesh.isWatertight)
+        XCTAssert(mesh.polygons.areWatertight)
+        XCTAssertEqual(mesh.polygons.count, 8)
+        XCTAssertEqual(mesh, .extrude(path, along: .square(), faces: .frontAndBack))
+    }
+
+    func testExtrudeOpenLineAlongOpenPath() {
+        let path = Path([
+            .point(0, 0),
+            .point(0, 1),
+        ])
+        let mesh = Mesh.extrude(path, along: path)
+        XCTAssert(mesh.isWatertight)
+        XCTAssert(mesh.polygons.areWatertight)
+        XCTAssertEqual(mesh.polygons.count, 2)
+        XCTAssertEqual(mesh, .extrude(path, along: path, faces: .frontAndBack))
+    }
+
+    func testLoftCircleToSquare() {
+        let shapes = [
+            Path.circle(),
+            Path.square().translated(by: Vector(0, 0, 1)),
+        ]
+
+        let loft = Mesh.loft(shapes)
+        XCTAssert(loft.isWatertight)
+        XCTAssert(loft.polygons.areWatertight)
+        XCTAssertEqual(loft.polygons.count, 22)
+    }
+
+    func testLoftSquareToCircle() {
+        let shapes = [
+            Path.square(),
+            Path.circle().translated(by: Vector(0, 0, 1)),
+        ]
+
+        let loft = Mesh.loft(shapes)
+        XCTAssert(loft.isWatertight)
+        XCTAssert(loft.polygons.areWatertight)
+        XCTAssertEqual(loft.polygons.count, 22)
+    }
+
+    func testLoftCircleToOpenPath() {
+        let shapes = [
+            Path.circle(),
+            Path([
+                .point(-1, -1, 1),
+                .point(1, -1, 1),
+                .point(1, 1, 1),
+                .point(-1, -1, 1),
+            ]),
+        ]
+
+        let loft = Mesh.loft(shapes)
+        XCTAssert(loft.isWatertight)
+        XCTAssert(loft.polygons.areWatertight)
+        XCTAssertEqual(loft.polygons.count, 21)
+    }
+
+    func testLoftOpenPathToCircle() {
+        let shapes = [
+            Path([
+                .point(-1, -1),
+                .point(1, -1),
+                .point(1, 1),
+                .point(-1, -1),
+            ]),
+            Path.circle().translated(by: Vector(0, 0, 1)),
+        ]
+
+        let loft = Mesh.loft(shapes)
+        XCTAssert(loft.isWatertight)
+        XCTAssert(loft.polygons.areWatertight)
+        XCTAssertEqual(loft.polygons.count, 21)
+    }
+
+    func testLoftEmptyPathToPath() {
+        let shapes = [
+            Path([]),
+            Path.square(),
+        ]
+
+        let loft = Mesh.loft(shapes)
+        XCTAssert(loft.isWatertight)
+        XCTAssert(loft.polygons.areWatertight)
+        XCTAssertEqual(loft.polygons.count, 2)
+    }
+
+    func testLoftPathToEmptyPath() {
+        let shapes = [
+            Path.square(),
+            Path([]),
+        ]
+
+        let loft = Mesh.loft(shapes)
+        XCTAssert(loft.isWatertight)
+        XCTAssert(loft.polygons.areWatertight)
+        XCTAssertEqual(loft.polygons.count, 2)
     }
 
     // MARK: Stroke
@@ -167,8 +347,21 @@ class MeshShapeTests: XCTestCase {
         XCTAssertEqual(mesh.polygons.count, 5)
     }
 
-    func testStrokeSqaureWithTriangle() {
+    func testStrokeSquareWithTriangle() {
         let mesh = Mesh.stroke(.square(), detail: 3)
         XCTAssertEqual(mesh.polygons.count, 12)
+    }
+
+    func testStrokePathWithCollinearPoints() {
+        let path = Path([
+            .point(0, 0),
+            .point(0.5, 0),
+            .point(0.5, 1),
+            .point(-0.5, 1),
+            .point(-0.5, 0),
+            .point(0, 0),
+        ])
+        let mesh = Mesh.stroke(path, detail: 3)
+        XCTAssertEqual(mesh.polygons.count, 15)
     }
 }