+++ Euclid/CHANGELOG.md
@@ -1,3 +1,32 @@
+## [0.5.25](https://github.com/nicklockwood/Euclid/releases/tag/0.5.25) (2022-07-03)
+
+- Added `Mesh.submeshes` property
+- Loft now supports path sections with differing point counts
+- Fixed normals at joints between multipart loft or extrusion shapes
+- Fixed crash when lofting empty paths
+
+## [0.5.24](https://github.com/nicklockwood/Euclid/releases/tag/0.5.24) (2022-06-26)
+
+- Fixed crash when extruding along collinear points
+- Fixed bug when lofting intermediate open paths
+- Improved makeWatertight() implementation
+- Improved vertex merging logic to only merge vertices that cause holes
+- Fixed bugs in triangulation logic that could introduce holes
+- Fixed case where mesh was incorrectly assumed to be watertight
+- Vector equality is now exact again, instead of using quantization internally
+- Fixed spurious assertion when using compound paths
+
+## [0.5.23](https://github.com/nicklockwood/Euclid/releases/tag/0.5.23) (2022-06-09)
+
+- Fixed crash in `mergingSimilarVertices()` function
+- Used heuristic for `mergingSimilarVertices()` instead of hard-coded epsilon
+- Increased epsilon precision again to fix cracking issues (and added tests)
+
+## [0.5.22](https://github.com/nicklockwood/Euclid/releases/tag/0.5.22) (2022-06-02)
+
+- Reduced epsilon precision to fix assertion when creating `Mesh` from `SCNGeometry`
+- Added WASM (Web Assembly) compatibility tests
+
 ## [0.5.21](https://github.com/nicklockwood/Euclid/releases/tag/0.5.21) (2022-05-28)
 
 - Fixed a glitch with how stroke or extrude-along was applied to paths with sharp corners